Bugfixes (#25)
* Build breaks on Raspberry Pi fixed
* Bugfix type detection, tA2lTypeId uin8_t, Warnings fixed
* Problem fixed, when atomic_uint_fast32_t is larger than 4 byte
* XCP_104.aml missing
* xcplite.c pow and math.h removed
* Math library missing for xcplib multi_thread_demo
* Examples use the simplified macros for DAQ

+++ xcplib/src/xcpEthTl.c
@@ -591,7 +591,7 @@ int32_t XcpTlHandleTransmitQueue(void) {
             mutexLock(&gXcpTl.CtrMutex);
             uint32_t lost;
             tQueueBuffer queueBuffer = QueuePeek(gXcpTl.Queue, flush, &lost);
-            gXcpTl.Ctr += lost; // Increase packet counter by lost packets
+            gXcpTl.Ctr += (uint16_t)lost; // Increase packet counter by lost packets
             uint16_t l = queueBuffer.size;
             const uint8_t *b = queueBuffer.buffer;
             if (b == NULL) {
